Multimodal improvements, and basic support for `meta-llama/Llama-3.2-11B-Vision-Instruct` (#771)
-- Support additional input features for Llama 3.2 Vision models in
`VisionLanguageCollatorWithPadding` and `VisionLanguageSftDataset`
-- Add custom sequence collation `torch_utils` based on
`torch.nn.utils.rnn.pad_sequence()`. More predictable, and works better
for complex shape features like `cross_attention_mask`
-- SFT tuning jobs for `meta-llama/Llama-3.2-11B-Vision-Instruct` run
w/o error with the changes
-- Fixed a bug when sequences are always padded to maximum model length
(not to the max length of a sequence in a batch), which makes trainings
faster,

Towards OPE-468, OPE-642, OPE-644

+++ scripts/benchmarks/minimal_multimodal_training.py
@@ -165,7 +165,7 @@ def test_multimodal_trainer(
     #
     model_params = ModelParams(
         model_name=model_name.value,
-        torch_dtype_str="float16",
+        torch_dtype_str="bfloat16",
         trust_remote_code=True,
         chat_template=_get_chat_template(model_name),
         freeze_layers=_get_freeze_layers(model_name),  # TODO: fix freeze + fsdp
